requirements.txt
numpy
pillow
chromadb
torchvision
opencv-python
opencv-headless
facenet-pytorch